Add empty value field in dropdown menu for suggested values
The dropdown menu for TemplateData suggested values (DropdownInputWidget)
displays the first value as "selected", even if no value is actually set
or if the set value is not part of the predefined set of suggested values,
which is misleading. Consequently, we prepend the value list with an empty
field to handle this.