Display only labels of TemplateData suggested values for classical dropdown
When only suggested values are accepted for a TemplateData parameter, a
classical dropdown menu is used. WIth this commit, if labels are defined,
they are displayed alone (values are hidden as they may be confusing).
Values are still shown for ComboBoxInputWidget (i.e. when not limited to
suggested values) as they can be typed.